move the controller into the API namespace
