+++ config/routes.rb
@@ -11,7 +11,7 @@ Rails.application.routes.draw do
   match '/api/openid_connect/token' => 'openid_connect/token#options', via: :options
   get '/api/openid_connect/userinfo' => 'openid_connect/user_info#show'
   post '/api/risc/security_events' => 'risc/security_events#create'
-  post '/api/irs_attempts_api/security_events' => 'irs_attempts_api#create'
+  post '/api/irs_attempts_api/security_events' => 'api/irs_attempts_api#create'
 
   # SAML secret rotation paths
   SamlEndpoint.suffixes.each do |suffix|
