+++ spec/controllers/api/irs_attempts_api_controller_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe IrsAttemptsApiController do
+RSpec.describe Api::IrsAttemptsApiController do
   before do
     stub_analytics
